+++ src/main/java/net/mcreator/energized/procedures/TankUpdateTickProcedure.java
@@ -24,7 +24,7 @@ public class TankUpdateTickProcedure {
 							.ifPresent(capability -> _retval.set(capability.getFluidInTank(tank).getAmount()));
 				return _retval.get();
 			}
-		}.getFluidTankLevel(world, new BlockPos((int) x, (int) y, (int) z), 1) >= 10 && new Object() {
+		}.getFluidTankLevel(world, new BlockPos(x, y, z), 1) >= 10 && new Object() {
 			public int getFluidTankLevel(LevelAccessor level, BlockPos pos, int tank) {
 				AtomicInteger _retval = new AtomicInteger(0);
 				BlockEntity _ent = level.getBlockEntity(pos);
@@ -33,7 +33,7 @@ public class TankUpdateTickProcedure {
 							.ifPresent(capability -> _retval.set(capability.getFluidInTank(tank).getAmount()));
 				return _retval.get();
 			}
-		}.getFluidTankLevel(world, new BlockPos((int) x, (int) (y - 1), (int) z), 1) <= new Object() {
+		}.getFluidTankLevel(world, new BlockPos(x, y - 1, z), 1) <= new Object() {
 			public int getFluidTankCapacity(LevelAccessor level, BlockPos pos, int tank) {
 				AtomicInteger _retval = new AtomicInteger(0);
 				BlockEntity _ent = level.getBlockEntity(pos);
@@ -42,7 +42,7 @@ public class TankUpdateTickProcedure {
 							.ifPresent(capability -> _retval.set(capability.getTankCapacity(tank)));
 				return _retval.get();
 			}
-		}.getFluidTankCapacity(world, new BlockPos((int) x, (int) (y - 1), (int) z), 1) - 10) {
+		}.getFluidTankCapacity(world, new BlockPos(x, y - 1, z), 1) - 10) {
 			fillValue = new Object() {
 				public int fillTankSimulate(LevelAccessor level, BlockPos pos, int amount) {
 					AtomicInteger _retval = new AtomicInteger(0);
@@ -52,16 +52,16 @@ public class TankUpdateTickProcedure {
 								capability -> _retval.set(capability.fill(new FluidStack(Fluids.LAVA, amount), IFluidHandler.FluidAction.SIMULATE)));
 					return _retval.get();
 				}
-			}.fillTankSimulate(world, new BlockPos((int) x, (int) (y - 1), (int) z), 10);
+			}.fillTankSimulate(world, new BlockPos(x, y - 1, z), 10);
 			{
-				BlockEntity _ent = world.getBlockEntity(new BlockPos((int) x, (int) y, (int) z));
+				BlockEntity _ent = world.getBlockEntity(new BlockPos(x, y, z));
 				int _amount = (int) fillValue;
 				if (_ent != null)
 					_ent.getCapability(CapabilityFluidHandler.FLUID_HANDLER_CAPABILITY, null)
 							.ifPresent(capability -> capability.drain(_amount, IFluidHandler.FluidAction.EXECUTE));
 			}
 			{
-				BlockEntity _ent = world.getBlockEntity(new BlockPos((int) x, (int) (y - 1), (int) z));
+				BlockEntity _ent = world.getBlockEntity(new BlockPos(x, y - 1, z));
 				int _amount = (int) fillValue;
 				if (_ent != null)
 					_ent.getCapability(CapabilityFluidHandler.FLUID_HANDLER_CAPABILITY, Direction.UP)
